requirements.txt
opencv-python-headless>=4.8.0,<5.0.0
numpy>=1.24.0,<2.0.0
Pillow>=10.0.0,<11.0.0
chromadb>=0.4.18,<0.5.0
fastapi>=0.109.0,<1.0.0
uvicorn>=0.27.0,<1.0.0
httpx>=0.26.0,<1.0.0
click>=8.1.0,<9.0.0
rich>=13.0.0,<14.0.0
python-dotenv>=1.0.0,<2.0.0
pydantic>=2.5.0,<3.0.0
gradio>=4.44.0,<5.0.0
pytest>=7.4.0,<8.0.0
pytest-asyncio>=0.23.0,<1.0.0
pytest-mock>=3.12.0,<4.0.0
